Avances para el método eliminar estudiante
