+++ shc-client/src/screens/coordinator/NewActivityScreen.js
@@ -175,8 +175,6 @@ const NewActivityScreen = ({ navigation }) => {
         />
       </View>
 
-
-
     </ScrollView>
